Update 1.1.0.0
Changes:
- Added the Workbench class, complete with code (unfinished)
- Changed Weapon.java to fit new class
- Changed Game.java to fit new class
- Changed Fight.java (fixes)

Overview:
- Added extra feature
- Arranged other classes to feature

Notes:
- Testing not done (tomorrow's job)
- Done at 11PM (not high quality)
- Comments to be added tomorrow.

+++ item/Weapon.java
@@ -2,7 +2,7 @@ package item;
 
 public class Weapon { // Creates a master class for all weapons
     public String name;
-    public int hitBonus;
+    public double hitBonus;
     public int baseDamage;
     public String damageType;
     public int bonusDamage;
